Remove old types and refactor
