+++ api/sonicapi/prepare_bundle.go
@@ -170,52 +170,6 @@ func flattenTransactions(args RPCExecutionProposal) ([]ethapi.TransactionArgs, m
 	return flatTxs, needGasLimit, needGasPrice, nil
 }
 
-// sanitizeBlockRange checks that the provided block range is valid and within
-// allowed limits, defaulting to a sensible range if not provided.
-func sanitizeBlockRange(currentBlock uint64, blockRange *RPCRange) (RPCRange, error) {
-	if blockRange == nil {
-		maxRange := bundle.MakeMaxRangeStartingAt(currentBlock + 1)
-		return RPCRange{
-			Earliest: hexutil.Uint64(maxRange.Earliest),
-			Latest:   hexutil.Uint64(maxRange.Latest),
-		}, nil
-	}
-
-	if blockRange.Latest == 0 && blockRange.Earliest != 0 {
-		blockRange.Latest = hexutil.Uint64(uint64(blockRange.Earliest) + bundle.MaxBlockRange - 1)
-		return RPCRange{
-			Earliest: blockRange.Earliest,
-			Latest:   blockRange.Latest,
-		}, nil
-	}
-
-	// earliest not specified but latest is specified, set earliest to latest - maxRange + 1, but not less than currentBlock + 1 and check overflow
-	if blockRange.Earliest == 0 && blockRange.Latest != 0 {
-		if blockRange.Latest < hexutil.Uint64(currentBlock+bundle.MaxBlockRange) {
-			blockRange.Earliest = hexutil.Uint64(currentBlock + 1)
-		} else {
-			return RPCRange{}, fmt.Errorf("invalid block range: latest block %d is too far in the future; earliest block must be at least %d", blockRange.Latest, currentBlock+1)
-		}
-	}
-
-	if blockRange.Latest < hexutil.Uint64(currentBlock) {
-		return RPCRange{}, fmt.Errorf("invalid block range: latest block %d is earlier than current block %d", blockRange.Latest, currentBlock)
-	}
-
-	if uint64(blockRange.Latest) < uint64(blockRange.Earliest) {
-		return RPCRange{}, fmt.Errorf("invalid block range: latest block %d is earlier than earliest block %d", blockRange.Latest, blockRange.Earliest)
-	}
-
-	if uint64(blockRange.Latest-blockRange.Earliest+1) > bundle.MaxBlockRange {
-		return RPCRange{}, fmt.Errorf("invalid block range: range %d is too large; must be at most %d blocks", blockRange.Latest-blockRange.Earliest+1, bundle.MaxBlockRange)
-	}
-
-	return RPCRange{
-		Earliest: blockRange.Earliest,
-		Latest:   blockRange.Latest,
-	}, nil
-}
-
 // injectPlanHashIntoAccessLists appends the BundleOnly access-list entry with planHash to every tx.
 func injectPlanHashIntoAccessLists(txs []ethapi.TransactionArgs, planHash common.Hash) {
 	for i, tx := range txs {
